% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhub.R
\name{bulk_rhub_for_lots_of_mails}
\alias{bulk_rhub_for_lots_of_mails}
\title{Check all rhub platforms}
\usage{
bulk_rhub_for_lots_of_mails()
}
\value{
rhub check_for_cran check_on_centos check_on_debian check_on_fedora check_on_linux check_on_macos check_on_ubuntu check_on_windows
}
\description{
Check all rhub platforms
}
