% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atlaunch.R
\name{fill_desc}
\alias{fill_desc}
\title{Fill description}
\usage{
fill_desc(name, Title, Description, repo)
}
\arguments{
\item{name}{name of the package}

\item{Title}{Title of the package}

\item{Description}{Desc of the package}

\item{repo}{Github repo of the package}
}
\value{
a modified DESCRIPTION
}
\description{
Fill description
}
